LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/scroll-timeline-progress.tentative-expected.txt


Harness Error (TIMEOUT), message = null

FAIL animation.progress reflects the progress of a scroll animation as a number between 0 and 1 assert_equals: The progress is null since the currentTime is unresolved. expected (object) null but got (undefined) undefined
TIMEOUT animation.progress reflects the overall progress of a scroll animation with multiple iterations. Test timed out
NOTRUN animation.progress reflects the overall progress of a scroll animation that uses a view-timeline.
NOTRUN progresss of a view-timeline is bounded between 0 and 1.
